# artifacts/api-server/src/routes/storage.ts
import { Router, type IRouter, type Request, type Response } from "express";
import { Readable } from "stream";
import {
  RequestUploadUrlBody,
  RequestUploadUrlResponse,
} from "@workspace/api-zod";
import { ObjectStorageService, ObjectNotFoundError } from "../lib/objectStorage";
import { requireAuth, type AuthenticatedRequest } from "../middlewares/requireAuth";
import { db, projectAttachmentsTable, taskAttachmentsTable, tasksTable, usersTable } from "@workspace/db";
import { eq, and } from "drizzle-orm";

const router: IRouter = Router();
const objectStorageService = new ObjectStorageService();

/**
 * POST /storage/uploads/request-url
 *
 * Request a presigned URL for file upload.
 * The client sends JSON metadata (name, size, contentType) — NOT the file.
 * Then uploads the file directly to the returned presigned URL.
 * Requires authentication to prevent anonymous storage abuse.
 */
router.post("/storage/uploads/request-url", requireAuth, async (req: Request, res: Response) => {
  const parsed = RequestUploadUrlBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: "Missing or invalid required fields" });
    return;
  }

  try {
    const { name, size, contentType, taskId, projectId } = parsed.data;

    // Organise uploads by task so each ticket has its own folder
    const subDir = taskId ? `uploads/tasks/${taskId}` : projectId ? `uploads/projects/${projectId}` : "uploads";
    const { uploadURL, objectPath } = await objectStorageService.getObjectEntityUploadURL(subDir);

    req.log.info({ objectPath }, "Generated upload URL");

    res.json(
      RequestUploadUrlResponse.parse({
        uploadURL,
        objectPath,
        metadata: { name, size, contentType },
      }),
    );
  } catch (error) {
    req.log.error({ err: error }, "Error generating upload URL");
    res.status(500).json({ error: "Failed to generate upload URL" });
  }
});

/**
 * GET /storage/public-objects/*
 *
 * Serve public assets from PUBLIC_OBJECT_SEARCH_PATHS.
 * These are unconditionally public — no authentication or ACL checks.
 * IMPORTANT: Always provide this endpoint when object storage is set up.
 */
router.get("/storage/public-objects/*filePath", async (req: Request, res: Response) => {
  try {
    const raw = req.params.filePath;
    const filePath = Array.isArray(raw) ? raw.join("/") : raw;
    const file = await objectStorageService.searchPublicObject(filePath);
    if (!file) {
      res.status(404).json({ error: "File not found" });
      return;
    }

    const response = await objectStorageService.downloadObject(file);

    res.status(response.status);
    response.headers.forEach((value, key) => res.setHeader(key, value));

    if (response.body) {
      const nodeStream = Readable.fromWeb(response.body as ReadableStream<Uint8Array>);
      nodeStream.pipe(res);
    } else {
      res.end();
    }
  } catch (error) {
    req.log.error({ err: error }, "Error serving public object");
    res.status(500).json({ error: "Failed to serve public object" });
  }
});

/**
 * GET /storage/objects/*
 *
 * Serve private object entities. Requires authentication.
 * Access is restricted to users who uploaded the attachment or who are admins.
 * The object path is matched against task_attachments to verify authorization.
 */
router.get("/storage/objects/*path", requireAuth, async (req: AuthenticatedRequest, res: Response) => {
  try {
    const raw = req.params.path;
    const wildcardPath = Array.isArray(raw) ? raw.join("/") : raw;
    const objectPath = `/objects/${wildcardPath}`;

    const clerkId = req.userId;
    if (!clerkId) {
      res.status(401).json({ error: "Unauthorized" });
      return;
    }

    const [dbUser] = await db.select({ id: usersTable.id, role: usersTable.role })
      .from(usersTable)
      .where(eq(usersTable.clerkId, clerkId));

    if (!dbUser) {
      res.status(401).json({ error: "Unauthorized" });
      return;
    }

    if (dbUser.role !== "admin") {
      const [projectAttachment] = await db
        .select({ uploadedById: projectAttachmentsTable.uploadedById })
        .from(projectAttachmentsTable)
        .where(eq(projectAttachmentsTable.objectPath, objectPath));

      if (!projectAttachment) {
        const [attachment] = await db
        .select({ uploadedById: taskAttachmentsTable.uploadedById, taskId: taskAttachmentsTable.taskId })
        .from(taskAttachmentsTable)
        .where(eq(taskAttachmentsTable.objectPath, objectPath));

        if (!attachment) {
          res.status(403).json({ error: "Forbidden" });
          return;
        }

        if (attachment.uploadedById !== dbUser.id) {
          const [task] = await db
            .select({ assigneeId: tasksTable.assigneeId, followerIds: tasksTable.followerIds })
            .from(tasksTable)
            .where(eq(tasksTable.id, attachment.taskId));

          const isAssignee = task?.assigneeId === dbUser.id;
          const isFollower = task?.followerIds?.includes(dbUser.id) ?? false;

          if (!isAssignee && !isFollower) {
            res.status(403).json({ error: "Forbidden" });
            return;
          }
        }
      }
    }

    const objectFile = await objectStorageService.getObjectEntityFile(objectPath);
    const response = await objectStorageService.downloadObject(objectFile);

    res.status(response.status);
    response.headers.forEach((value, key) => res.setHeader(key, value));

    if (response.body) {
      const nodeStream = Readable.fromWeb(response.body as ReadableStream<Uint8Array>);
      nodeStream.pipe(res);
    } else {
      res.end();
    }
  } catch (error) {
    if (error instanceof ObjectNotFoundError) {
      req.log.warn({ err: error }, "Object not found");
      res.status(404).json({ error: "Object not found" });
      return;
    }
    req.log.error({ err: error }, "Error serving object");
    res.status(500).json({ error: "Failed to serve object" });
  }
});

export default router;